add initial ethernet remote transfer stress test infra
This test infrastructure was developed with the intention of better stress testing the ethernet remote transfer code on host and in erisc FW with the aim of exposing bugs in a smaller, more isolated, and better controlled test environment than during large real-world workloads.

Currently the infra does not support readback to check for correctness. So far we only care about hangs. Readback and checking will be added in the future.  Additionally, more test configurations will be added with different transfer size, type, and destination distributions.

Add test fixtures that will exit out if the present hardware is not intended for the current test
 - Conditionally skip tests that are unrunnable on the current machine (due to incorrect number of chips for example)

Add ability for command trace to dump the raw C++ code which can then be used to directly generate a test case

Below is an example test with a custom write size distribution specified:
```
    auto write_size_generator = ConstrainedTemplateTemplateGenerator<transfer_size_t, double, std::normal_distribution>(
        seed, std::normal_distribution<>(10000, 3000), [](double x) -> transfer_size_t { return size_aligner_32B(static_cast<transfer_size_t>((x >= 4) ? x : 4)); });

    auto dest_generator = get_default_full_dram_dest_generator(seed, device.get());
    auto address_generator = get_default_address_generator(seed, 0x100000, 0x5000000);

    try {
        RunMixedTransfers(
            *device,
            10000,
            0,

            transfer_type_weights_t{.write = 1., .rolled_write = 0., .read = 0., .epoch_cmd_write = 0.},

            WriteCommandGenerator(dest_generator, address_generator, write_size_generator),
            build_dummy_rolled_write_command_generator(*device),
            build_dummy_write_epoch_cmd_command_generator(*device),
            build_dummy_read_command_generator(*device),

            false, // Set to true if you want to emit the command history code to command line
            &command_history
        );
    } catch (...) {
        print_command_history_executable_code(command_history);
    }
```