Bump DMA read threshold to 128
In BBE tt_log_server can be enabled to monitor for log being placed
by RISK cores. This is a background thread that just keeps running and
reading from "mailbox" on each risk on each tensix.
Mailbox size is 64bytes which triggers the DMA read and as it keeps
running in backtround ti can be concurrent read to any other DMA read.
Workaround this issue by bumping threshold for DMA read from 32->128.