Add API for fast writes to static TLBs
Existing code paths for performing such a write suffer from limitations:
The "REG_TLB" path reconfigures a single 16M TLB and performs multiple
hash table lookups (costing on the order of hundreds of nanoseconds
each) prior to performing the write (on the order of tens of
nanoseconds). The non "REG_TLB" path can take two directions: use of an
alternate 16M TLB (incurring the same performance penalty) or use of a
statically-mapped TLB (if such a mapping exists). These paths contain
read-modify-write logic which may result in unexpected system behavior.

This change adds an API where the cost for hash table lookups is paid
up-front. The intent is for callers requiring fast write access to a
specific core to create and maintain a Writer object using this new API.