No ARCH_NAME in CI building umd_device (#88)
A continuation to #87 , to reflect the current state in our CI.
Related to #86 and a part of #51 .

There is now enough difference between building umd_device and umd_tests
to warrant separate files, since umd_tests still requires ARCH_NAME and
umd_device does not. This way it looks clearer.