Fix Grayskull tests (#138)
Simulated harvesting was hardcoded but it is expected to cover real
harvesting mask, which means it won't work on each card. This PR fixes
this.

For grayskull only, add real harvesting mask to the simulated one. This
will be fixed after the redesign naturally, since harvesting info will
come from tt::umd::ClusterDescriptor, so you could join it with
simulated one in test instead.