Allow GCC 9.4
This change allows UMD to be built and debugged with the
distribution-provided tools under Ubuntu 20.04.  This requires
CMAKE_C_COMPILER and CMAKE_CXX_COMPILER environment variables to be set
to gcc and g++, respectively.  Without them, the build will still
default to using a third-party compiler which the user must have
installed.