ClusterDescriptor API test - not supporting multiple clusters (#165)
A continuation on working on #154 
The first test showcases the basic usage of cluster descriptor.

The second test tests whether ClusterDescriptor can manage multiple
clusters of cards, cluster meaning connected chips (through ethernet).
It actually shows explicitly that this is not currently supported. So
this test fails on a system with multiple unconnected clusters of cards
(simple 2xN150 setup would suffice). We don't have a machine like that
in UMD pool, so the test passes on CI.