API test - simple IO (#159)
This change adds minimal setup to read/write including remote chips.
I've tested this on a setup with 3 N300 cards.

Motivation was that I wanted to change wait_for_non_mmio_flush to
include chip_id. Then I wanted to write a test. Then it didn't make
sense writing a test for this if basic IO is not setup.

This test adds to the collection of API tests which we should work on
reducing. Related to #154 . Related to #157

Minor changes:
- get_soc_descriptor had two different definitions, consolidated those.
- added is_chip_remote for convenience