Pull request template (#268)
### Ticket
No issue for this change

### Description
After discussion with @pjanevskiTT on how to handle API breaking UMD
PRs, we agreed to try with this approach. The point is that there should
be minimal time between merging UMD PR and client (tt_metal and
tt_debuda) PRs. This should then not block other improvements/PRs.

### List of the changes
- Added pull_request_template file.

### API Changes
There are no API changes in this PR.