Print hugepage 0 size warning only if mapping fails (#278)
### Issue
Fixes https://github.com/tenstorrent/tt-umd/issues/275

### Description
Aftermath of #168 . Added a warning by design. This was being printed
even if everything worked correctly, but hugepage file was created for
the first time.

### List of the changes
- Moved warning not to printed everytime but only in case of mapping
failure.
- Minor PR template addition to motivate people to comment on testing

### Testing
Manually tested on t3008 machine that if no hugepages are available this
warning is printed. Also tested this warning is not printed if
everything works but is first hugepage setup.

### API Changes
There are no API changes in this PR.