Refactor our CI yamls (#319)
### Issue
No current issue, related to already closed #45 

### Description
As we're using CI for some time, I found that it actually brings
unnecessary complications to have a layer of yamls just designating what
is done on pr, optionally on pr, and on push. This can be directly added
to each of the jobs.

Another effect is that in the "Actions" tab, we won't have On PR,
Optional On PR, and On Push where each includes several different jobs.
We will now have Build device, build and run tests, and pre-commit ,
which will run both on PR branches and main.

### List of the changes
- Created a build-and-run-all-tests.yml which wraps build-tests and
run-tests for all archs. This just nicely corresponds exactly to
previous on-pr-opt.
- Deleted on-pr, on-pr-opt and on-push ymls
- Added to all referenced .ymls the config to run on pr and/or on push,
so that same jobs are called on same triggers.
- Added timeout parameter to a couple of jobs.
- I will change in settings which jobs are mandatory, if they are now
named slightly different.

### Testing
CI passes on this PR

### API Changes
There are no API changes in this PR.