Remove setup_core_to_tlb_map (#403)
### Issue
Related to #44 but also on a path of #351 

### Description
This function in API is a surplus. All the info is already available to
the Cluster class through configure_tlb calls.
This is initial PR which changes the API. The following PR won't change
the API but will restructure functions internally to introduce
TLBManager class.

### List of the changes
- Removed setup_core_to_tlb_map. Fill up the same internal structure
through configure_tlb api
- tlbs_init_per_chip is also unnecessary
- Extracted the logic for checking whether static tlb is setup

### Testing
Covered through existing tests.

### API Changes
This PR has API changes:
- [x] tt_metal approved PR pointing to this branch:
https://github.com/tenstorrent/tt-metal/pull/16048
- [X] tt_lens doesn't use this API call.