Remove describe_tlb function (#419)
### Issue
Related to #417 

### Description
Seems like we've had a redundant get_tlb_configuration and describe_tlb
functions. I also noted that sometimes we downcasted 64bit address to
32bit, so this should solve those problems if they were even present.

### List of the changes
- Removed describe_tlb everywhere
- Changed describe_tlb calls with get_tlb_configuration
- API now returns tlb_configuration type.

### Testing
Existing CI tests

### API Changes
This PR has API changes. I've left old API in place so that this change
is not breaking, but is marked as deprecated.