Fix wormhole ETH translation (#469)
### Issue
While working on https://github.com/tenstorrent/tt-umd/issues/439

### Description
Translation was being calculated by using logical coordinates. But we
want logical coordinates to be in the same order as defined in the input
yaml. Since these are not consecutive (meaning consecutive logical
coords are not consecutive any other coords), this breaks translation
logic.
BH to be fixed.

### List of the changes
- Fixed the wh eth translation test not to have any logic
- Fixed eth translation to match the earlier one from cluster

### Testing
Fixed the test to have hardcoded values.

### API Changes
There are no API changes in this PR.