Simple hack for #526 until it is properly fixed (#534)
Exposing global variable for using noc1 in tlb.
Disabling coordinate translation when setting up tlb for noc1.

Note: this is bad implementation since required coordinates for
communication with UMD are virtual which get converted into physical,
but in tt-lens we are converting noc1 into virtual as if it is noc0, UMD
reverses this so on device ends up noc1 coordinate...