Serialize cluster into yaml (#499)
### Issue

Solving part of #431 

### Description

Add a function to serialize cluster to the yaml (reverse of what we were
doing so far). Add flag to all Cluster constructors to create mock chips
inside Cluster if user wants that. Expose this (de)serializing through
the API as mentioned in the issue above. Output yaml look like (note the
styling difference than original, but in YAML world it is the same)

```
arch:
  1: wormhole_b0
  0: wormhole_b0
ethernet_connections:
  -
    - chip: 1
      chan: 1
    - chip: 0
      chan: 9
  -
    - chip: 1
      chan: 0
    - chip: 0
      chan: 8
chips_with_mmio:
  - 0: 0
harvesting:
  1:
    noc_translation: true
    harvest_mask: 513
  0:
    noc_translation: true
    harvest_mask: 129
boardtype:
  1: n300
  0: n300 
```

- Add serialize function to Cluster
- Add serialize function cluste descriptor 
- Add flag to all Cluster constructors to create mock chips