Fix UMD unit tests on 80.15 fw bundle (#543)
### Issue

UMD unit tests were failing on new firmware release 80.15.

### Description

Resize the telemetry table tag array to fit all the new tags in the
firmware. The new tags are out of the original spec so resizing to 64 in
order to have room for some additional if needed. Follow up issue to
size this properly or just support any tag value is #544

### List of the changes

- Resize the telemetry tag table

### Testing

Flashed the new fw and ran CI

### API Changes
/