Hacks for DMA support on SiFive P550 Premier
Use a DMA buffer allocated by KMD rather than allocating a buffer and
asking the driver to pin/map it.  This change requires a KMD that has
been modified to support 1G DMA buffers (raised from 256M).

Memory allocated by the application and pinned/IOMMU mapped by the
driver is not coherent on this platform and requires cache maintenance
operations to be correct.

Memory allocated by the driver with dma_alloc_coherent does not require
cache maintenance operations because it uses page protection flags with
_PAGE_UNCACHE set in the page table entries.  On this platform, the
vendor kernel has explicit logic handling this case:

https://github.com/sifive/riscv-linux/blob/c25a91c16ef256866e083b7f67fa96bb35e4a529/arch/riscv/include/asm/pgtable.h#L403-L407

The SOC TRM:
https://github.com/eswincomputing/EIC7700X-SoC-Technical-Reference-Manual/releases/tag/v1.0.0-20250103

Although the SOC TRM suggests that it is DMA coherent (see 3.4.1.1), my
testing indicates that it is not.