Blackhole ETH id mapping fix (#599)
### Issue

#588 ETH id from ETH core is not considering harvesting, and we would
want to have logical ETH channels which consider harvesting in cluster
descriptor. We put ETH id from the core in the cluster descriptor, so
now considering harvesting as well

### Description

Add mapping of ETH id from the core to the logical ETH channel to know
which logical channel to put in the cluster descriptor instead od ETH id
from ETH core.

### List of the changes

- Add mapping of ETH id from ETH core to logical eth channel
- Add logical eth channels to cluster descriptor
- Add verification that there is no higher ETH logical channel that
number of channels

### Testing
CI + ran locally to check the mapping

### API Changes
/