Remove assert for Blackhole arc messages (#615)
### Issue

Remove assert for Blackhole ARC messages to enable custom use of arc_msg
interface. This path is not used by UMD so we didn't see this, but
someone would want to use this interface in a custom way. tt-exalens
test was failing

### Description

Remove assert for BH

### List of the changes

- Remove assert for BH

### Testing

CI + tt-exalens tests pass

### API Changes
/