noc_translation_enabled uninitialized fix (#632)
### Issue
This arised as an issue in
https://github.com/tenstorrent/tt-umd/pull/626

### Description
Setting this variable was silently removed in
https://github.com/tenstorrent/tt-umd/pull/609, but it was still read
across the code. Reading uninitialized variable gives undefined value.
What's interesting is that printing this value gave "32" even though it
is a bool.
I was previously aware that we have two ways of reading
noc_translation_enabled, one through cluster_descriptor (or during
cluster topology discovery), and another through soc_descriptor, but
haven't consolidated them. Referenced PR removed this reading from
soc_descriptor, because it is not even present at any soc descriptor, so
this change was fine.

### List of the changes
- Removes noc_translation_id_enabled
- Move usages to noc_translation_enabled

### Testing
For some strange reason, only
https://github.com/tenstorrent/tt-umd/pull/626 surfaced this issue. I
ran the test before this change and after the change, manually, and the
failure doesn't appear anymore. Note that the failure is not
deterministic, you have to run it ~10 times to check if it will fail
anytime or not.

### API Changes
There are no API changes in this PR.