Implement ARC telemetry reader class (#661)
### Issue

#508 

### Description

Implement telemetry reader class. This class should hide the details of
reading the telemetry on Wormhole and Blackhole. It provides simple API
to check whether some tag is present in the telemetry and read the
appropriate tag value.

### List of the changes

- Implement base telemetry reader class
- Derive wh and bh telemetry readers and move logic to these classes
- Use telemetry reader class wherever it is needed to read the telemetry
- Add tags for Wormhole telemetry

### Testing

CI. A lot of tests are going through telemetry to get basic info about
the device so that is a good test in itself. Additional tests added for
each architecture

### API Changes
/