Wait for Wormhole DRAM training (#672)
Wait for DRAM training on wormhole LocalChip to finish.

Creating chip should wait on DRAM to finish training. There is a push
for Blackhole as well #542, that is waiting for Blackhole fw release.
This PR implements wait for DRAM training on wormhole. This is part of
the push to have create-ethernet-map implemented in source.

- Add wait for DRAM training status on LocalChip creation
- Add possible status of WH dram training

CI + tested locally readout of the DRAM training status

/