TTDevice get max and min AICLK (#709)
### Issue

AICLK go busy value is not always going to be the same, but we can read
the maximum possible value on the device. So instead of waiting for 1000
on Wormhole, wait on the maximum for the specified device.

Aftermath of https://github.com/tenstorrent/tt-umd/pull/663 and
https://github.com/tenstorrent/tt-umd/pull/698

### Description

Add functions to get maximum and minimum possible AICLK from TTDevice.
Wait on those values when changing power state.

### List of the changes

- Add function for getting maximum AICLK from device
- Add function for getting minimum AICLK from device
- Wait on those values when setting power state
- Change test for Wormhole since we can't hardcode value now
- Fix remote arc message timeout

### Testing
CI

### API Changes
/

---------

Co-authored-by: Bojan Rosko <brosko@tenstorrent.com>