Pass a custom file to serialize method (#694)
### Issue
By a request from this thread
https://github.com/tenstorrent/tt-metal/pull/19862

### Description
Currently our serializing API returns the filepath, but this change now
allows it to accept the filepath at which the clusterdescriptor will be
written to.

### List of the changes
- Added get_default_cluster_descriptor_file_path to
tt_ClusterDescriptor, and moved the filename generation part of
serialization there.
- The Cluster::serialize_to_file can now accept the filepath.
- Other changes to allow for this change

### Testing
No testing

### API Changes
There is an API changes in this PR, but shouldn't be breaking since the
arg has a default value.

---------

Co-authored-by: Artem Yerofieiev <169092593+ayerofieiev-tt@users.noreply.github.com>