Add read/write reg functions, removing fallback_tlbs (#738)
### Issue
On a path to #248 , but #737

### Description
We want to remove fallback_tlbs from the API, and have this
internalised. As a first step, give out the new _reg API and for the
existing one provide default fallback_tlb_versions. Next step is clients
to switch to the new API calls, then I'll get back to this and remove it
here

### List of the changes
- Add write_to_device_reg and read_from_device_reg
- Set default fallback_tlb in write_to_device and read_from_device

### Testing
No testing

### API Changes
There are API changes in this PR, but they are only adding new options