Move read_from_non_mmio_device to RemoteChip (#682)
### Issue
On a path of #248

### Description
In this PR, remote read communication is moved to RemoteChip.
There's work on:
1. Initial changes to accommodate RemoteChip to even do what's needed
2. Actual move of read and write to RemoteChip
3. Additional complexities due to a fact that remote write can be done
on a LocalChip in case of ethernet broadcast.

1+2+3 seemed too much for a PR, so instead I did just 1+2 by moving only
the read part, and 2+3 will be done in a next part where I'll move the
write part.

### List of the changes
- RemoteCommunication now accepts LocalChip instead of TTDevice
- Copied the read code from cluster.cpp instead of existing
RemoteCommunication's hacked read code
- In RemoteCommunication::write_non_mmio I did slight changes due to now
having LocalChip instead of TTDevice and LockManager.
- It only made sense that get_local_chip and get_remote_chip actually
return the right classes. I tried other options but I liked this one the
most. In any case, those two calls should not be used a lot (in the
final code where everywthing is moved to chip classes). Added
get_remote_chip call.
- RemoteChip can now do a read through read_from_device, same api as is
for LocalChip. This will get unified in Cluster in some of the following
PRs. Also, fallback_tlbs will be removed in some of the following PRs
- Slightly changed the order of arguments in read_non_mmio so that first
the target is given, then addresses, then size. Note that eth core used
is not passed anymore, but it is fetched from the local_chip_. This
change was actually done in
https://github.com/tenstorrent/tt-umd/pull/667

### Testing
Existing CI tests.
Also tt_metal tests, since I'm uncertain on the scope of our remote
tests:

### API Changes
There are no API changes in this PR.