Implement PCIE harvesting (#742)
### Issue

#700 

### Description

Implement PCIE harvesting in general. This is needed for Blackhole, but
so far we have managed this based on the board type. This PR introduces
relying on information from the device on PCIE port config. Even if this
information should be 1-1 with board type, it is more reliable to trust
the device configuration.

### List of the changes

- Add pcie harvesting mask
- Add pcie harvesting in coordinate manager
- Remove fuzzy logic based on board type for blackhole
- Update tests
- Fix some tests to pass proper PCIE harvesting mask to structures

### Testing
CI

### API Changes
/