Fix PCI harvesting for P100 board (#772)
### Issue

#769 

### Description

We have two types of P100 cards. They use different PCI cores on
Blackhole so we can't rely on the check that used PCI core is always
going to be the same. This PR removes the check from soc descriptor and
changes some tests.

### List of the changes

- Remove the check for harvested PCI core on P100
- Fix reverse logic for PCI harvesting in coordinate manager
- Change the tests to test the change

### Testing
CI. Change was manually tested on P100 and P100A boards to verify that
tests pass.

### API Changes
/