Implement create ethernet map (#722)
### Issue

#73 #625. This PR is not going to close any of these issues, just a big
part in working towards closing it.

### Description

Implement topology discovery class for Wormhole with old routing fw.
This class should have the same functionality as CEM lib from luwen. The
idea is to replace CEM lib with our own implementation. In order to do
so, we need to secure wait on ARC, ETH and DRAM, same as we did for
Blackhole. #723 #724

Blackhole and 6U topology discovery should be moved to this class as
well. I just wanted to keep the scope of this PR smaller.

### List of the changes

- Implement topology discovery for Wormhole with old routing fw
- Remote get board type
- Remote arc msg
- Add test for topology discovery to run it on wormhole

### Testing

Topology discovery is still not integrated into main UMD code path.
Manually tested on N150, N300, T3K and TG configs.

### API Changes
/