Fix default translation for TRANSLATED BH coords (#783)
### Issue
No issue

### Description
Surfaced through https://github.com/tenstorrent/tt-umd/pull/747
Default translation is wrong for tensix for blackhole when noc
translation is not enabled.

### List of the changes
- make get_grid_size pure virtual
- change default translate tensix implementation to include bh case when
columns are harvested.

### Testing
We're missing translation tests for when noc translation is not enabled.
Manually ran this code on 747 change

### API Changes
There are no API changes in this PR.