Disable Blackhole NOC1 node id test (#787)
### Issue

Mitigation of #776 

### Description

Disable NOC1 node id test in case we need to bump UMD for 80.18 firmware
release since the test fails on fw 80.18. If the test is fixed in the
meantime, this PR can be closed. NOC1 working is not blocking tt-metal.
However, we would like to enable this back as fast as possible.

### List of the changes

- Disable NOC1 node id test

### Testing

CI

### API Changes
/