Wait for Blackhole DRAM training (#542)
### Issue

#478 

Note: the PR is probably going to timeout on Blackhole CI until the fw
from the syseng is released, so reviewers can do the review without me
checking this in, to be ready when the fw is released.

### Description

Implement waiting for DRAM training on Blackhole. Format of the data
provided by syseng to sw is described in
https://tenstorrent.atlassian.net/browse/SYS-1241. This PR integrates
the wait in our code according to the proposed ticket design.

### List of the changes

- Add function to wait for DRAM cores
- Implement the proposed design from the ticket
- Add functions to TTDevice that expose DRAM training info

### Testing
CI

### API Changes
/

---------

Co-authored-by: Bojan Roško <156314064+broskoTT@users.noreply.github.com>