Remove unused Chip APIs (#801)
### Issue
On a path of #250 

### Description
Remove some unused APIs from Chip to make the transition easier.

### List of the changes
- Remove acquire_mutex from Chip API, as it is not needed in Cluster
anymore
- Remove remote transfer eth core related Chip API, as it is not needed
in Cluster anymore
- Move read/write _reg functions to Chip with the default implementation
of using the non _reg variants. This was formerly in RemoteChip

### Testing
Client CIs

### API Changes
There are API changes in the whole PR stack which ends with :
https://github.com/tenstorrent/tt-umd/pull/800
I'll merge the whole stack once the whole stack is approved, and then
merge the changes in the clients.
Relevant links to CIs and PRs in clients are in the referenced PR