Introduce close device to Chip (#802)
### Issue
On a path of #250 

### Description
Introduce close_device() which will gain more content in the following
period.

### List of the changes
- Add close_device to chip.h
- Implement it trivially in derived classes

### Testing
Client CIs

### API Changes
There are API changes in the whole PR stack which ends with :
https://github.com/tenstorrent/tt-umd/pull/800
I'll merge the whole stack once the whole stack is approved, and then
merge the changes in the clients.
Relevant links to CIs and PRs in clients are in the referenced PR