Introduce send_tensix_risc_reset to Chip (#803)
### Issue
On a path of #250 

### Description
Introduce an API which sends risc_reset signal to all cores on a chip.
Motivation for this change was that before this change, if we tried to
use Mock/Simulation chip, the code would still go through the flow of
trying to setup ethernet broadcast for broadcasting tensix reset, and
then collapse to write_to_device_reg in case broadcast is not enabled.
This change makes this happen a bit earlier, at the beginning of the
broadcast_tensix_risc_reset_to_cluster function

### List of the changes
- Introduce send_tensix_risc_reset, which sends to all cores.
- Call send_tensix_risc_reset if broadcast is not enabled in
broadcast_tensix_risc_reset_to_cluster

### Testing
Client CIs

### API Changes
There are API changes in the whole PR stack which ends with :
https://github.com/tenstorrent/tt-umd/pull/800
I'll merge the whole stack once the whole stack is approved, and then
merge the changes in the clients.
Relevant links to CIs and PRs in clients are in the referenced PR