Simplify simulated_harvesting_mask usage (#804)
### Issue
On a path of #250 

### Description
Simplify simulated_harvesting_mask usage

### List of the changes
- Change some functions from td::unordered_map<chip_id_t,
HarvestingMasks>& simulated_harvesting_masks to HarvestingMasks
simulated_harvesting_masks, which refer to a single chip.
- get_*_harvesting_mask functions are now too small and bring more
confusion than clarity, so I removed them.
- perform_harvesting check is moved to get_harvesting_masks
- Check if harvesting_masks exist for a chip is moved to constructors

### Testing
Client CIs

### API Changes
There are API changes in the whole PR stack which ends with :
https://github.com/tenstorrent/tt-umd/pull/800
I'll merge the whole stack once the whole stack is approved, and then
merge the changes in the clients.
Relevant links to CIs and PRs in clients are in the referenced PR