Implement new constructor accepting ClusterOptions (#805)
### Issue
On a path of #250 

### Description
Introduce new Cluster constructor, which accepts ClusterOptions.
Motivation was that our number of constructors was exploding. This one
allows changing any one of the options in isolation, or in any
combination.

### List of the changes
- Introduce ChipType enum for chip_type argument, so we know which kind
of cluster to create.
- Introduce ClusterOptions which should provide options to the
constructor
- Changed create_mock_chip usage to go through ChipType
- New construct_soc_descriptor which seems like a better way to separate
responsibilities. The second two construct_chip_from_cluster are marked
as deprecated (removed in following PRs)
- Implement the new constructor, which should also service Chip and
Simulation chips now.
- get_chips_local_first now also verifies chips are present in the
cluster.
- target_devices in ClusterOptions are now unordered_set compared to a
set in other constructors
- Changed all tests to use the new ClusterOptions constructor. Note that
many of the values are default ones, and this is cleaned up in one of
the following PRs

### Testing
Client CIs

### API Changes
There are API changes in the whole PR stack which ends with :
https://github.com/tenstorrent/tt-umd/pull/800
I'll merge the whole stack once the whole stack is approved, and then
merge the changes in the clients.
Relevant links to CIs and PRs in clients are in the referenced PR