Simplify tests using default constructor values (#806)
### Issue
On a path of #250 

### Description
Simplify tests, take advantage of default values.

### List of the changes
- Changed much code in tests.
- Removed .target_devices argument whenever whole cluster was expected
- Removed num_channels and perform_harvesting when they have default
values.
- Changed DifferentConstructors tests a bit, added a new option

### Testing
Client CIs

### API Changes
There are API changes in the whole PR stack which ends with :
https://github.com/tenstorrent/tt-umd/pull/800
I'll merge the whole stack once the whole stack is approved, and then
merge the changes in the clients.
Relevant links to CIs and PRs in clients are in the referenced PR