Derive SimulationDevice from Chip (#807)
### Issue
On a path of #250 
Solves #249

### Description
Implement SimulationDevice as Chip as oppossed to tt_device.
Simulation should be created from the Cluster from now on.

### List of the changes
- Change : public tt_device to : public Chip
- Introduce simulator_directory option in ClusterOptions and propagate
it.
- Change SimulationDevice header so that it matches what's in the Chip.
- Change slightly implementation in SimulationDevice so it matches the
new API, the implementation is not actually changed that much.
- Change simulation tests since the API changed.

### Testing
Client CIs

### API Changes
There are API changes in the whole PR stack which ends with :
https://github.com/tenstorrent/tt-umd/pull/800
I'll merge the whole stack once the whole stack is approved, and then
merge the changes in the clients.
Relevant links to CIs and PRs in clients are in the referenced PR