Remove serialize from Cluster (#809)
### Issue
On a path of #250 

### Description
serialize in Cluster is unnecessary.

### List of the changes
- Remove serialize and serialize_to_file from Cluster
- Change the tests accordingly
- Slightly changed tt_ClusterDescriptor::serialize_to_file in case of no
path provided

### Testing
Client CIs

### API Changes
There are API changes in the whole PR stack which ends with :
https://github.com/tenstorrent/tt-umd/pull/800
I'll merge the whole stack once the whole stack is approved, and then
merge the changes in the clients.
Relevant links to CIs and PRs in clients are in the referenced PR