Remove MockupDevice (#811)
### Issue
On a path of #250 

### Description
ChipType::MOCK , or MockChip should be used instead.

### List of the changes
- Remove MockupDevice class

### Testing
Client CIs

### API Changes
There are API changes in the whole PR stack which ends with :
https://github.com/tenstorrent/tt-umd/pull/838
I'll merge the whole stack once the whole stack is approved, and then
merge the changes in the clients.
Relevant links to CIs and PRs in clients are in the referenced PR