Non silicon chips are local chips (#835)
### Issue
On a path of #250 

### Description
Simulation chips have been mistakenly put into remote chips, which made
tests fail.

### List of the changes
- Change bool create_mock_chips to ChipType chip_type
- Set simulator and mock chips as local chips.

### Testing
Client CIs

### API Changes
There are API changes in the whole PR stack which ends with :
https://github.com/tenstorrent/tt-umd/pull/838
I'll merge the whole stack once the whole stack is approved, and then
merge the changes in the clients.
Relevant links to CIs and PRs in clients are in the referenced PR