Move several minor functions to Chip class (#839)
### Issue
On a path of #250 

### Description
Move several minor function's implementation to Chip class.
Motivation was that code in cluster.cpp was assuming that local chips
had tt_device and sysmem_manager, which is not true for Simulation chip.
Moving this to Chip allows for custom implementation.

### List of the changes
- Move get_num_host_channels, get_host_channel_size,
get_fast_pcie_static_tlb_write_callable, get_clock, get_numa_node to
Chip
-  Removed this from Cluster
- All implementations were very simple, so not much changes, just moving
code around.
- Keep LocalChip in Remote to be able to call its functions

### Testing
Client CIs

### API Changes
There are API changes in the whole PR stack which ends with :
https://github.com/tenstorrent/tt-umd/pull/838
I'll merge the whole stack once the whole stack is approved, and then
merge the changes in the clients.
Relevant links to CIs and PRs in clients are in the referenced PR