Move deassert_risc_resets to Chip class (#840)
### Issue
On a path of #250 

### Description
Move deassert_risc_resets implementation to Chip class.
Motivation was that code in cluster.cpp was assuming that local chips
had tt_device and sysmem_manager, which is not true for Simulation chip.
Moving this to Chip allows for custom implementation.
Minor fix for eth_fw_version

### List of the changes
- Move deassert_risc_resets to Chip.
- Implementation for both local and remote chips turned out to be the
same, just the remote chip was calling it's local chip function.
- Allow eth_fw_version to be empty. We didn't have a wormhole chip
without eth previously, now we have through Simulation

### Testing
Client CIs

### API Changes
There are API changes in the whole PR stack which ends with :
https://github.com/tenstorrent/tt-umd/pull/838
I'll merge the whole stack once the whole stack is approved, and then
merge the changes in the clients.
Relevant links to CIs and PRs in clients are in the referenced PR