Add ubench tests for TLB read/write (#837)
### Issue

Part of the work done for #751 

### Description

Add ubench tests for TLB that were written to measure the perf as part
of the report for the issue. There are tests for read/writes using both
dynamic and static TLB for tensix/DRAM. nanobench infra that was there
for measuring is deleted, but there is a still dependency, we might want
to use this in the future. I think there is no point in adding
additional tests for TLBs, I think these showcase perf for all cases.
Since perf of TLB path is not something that can be used in metal's perf
path, I think this set of tests is enough for the following period.

### List of the changes

- Add test for dynamic TLB r/w tensix
- Add test for dynamic TLB r/w DRAM
- Add test for static TLB r/w tensix
- Add test for static TLB r/w DRAM

### Testing

Ran tests manually

### API Changes
/