WH: Don't map top 256MB of 4th memory channel (#875)
### Issue
N/A

### Description
Don't pin memory that isn't accessible by hardware anyway. This is in
anticipation of a KMD change that introduces ATU resource management
(step 1 for full WH IOMMU support).

### List of the changes
* Rehome `HUGEPAGE_CHANNEL_3_SIZE_LIMIT` constant so it can be used in
sysmem_manager.cpp and elsewhere.
* Respect this limit when doing ATU configuration for WH (existing code;
unchanged) _and_ pinning pages (added in this change).
* Remove comments that will become obsolete.

### Testing
Manual

### API Changes
There are no API changes in this PR. The memory in question wasn't
accessible to the device prior to this change.